test(mastracode-web): colocate storage domain tests
Move each Postgres storage suite beside the domain it exercises so ownership and future maintenance stay local to the implementation.\n\nCo-Authored-By: Mastra Code (openai/gpt-5.6-sol) <noreply@mastra.ai>
